% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/one.minus.ratio.100.R
\name{one.minus.ratio.100}
\alias{one.minus.ratio.100}
\title{Helper function for TRA and TBA}
\usage{
one.minus.ratio.100(num, den)
}
\arguments{
\item{num}{the numerator in 100*(1 - num/den)}

\item{den}{the numerator in 100*(1 - num/den)}
}
\description{
TRA and TBA both have the same form: 100*(1 - num/den).
Just specify num and den and get the desired calculation.
}
\examples{
one.minus.ratio.100(prevalence(rho_in=2.7, muci_in=10, theta_in=1.75),
                    prevalence(muci_in=10, theta_in=1.75))
}
\keyword{pim}
\keyword{pip}
\keyword{tba}
\keyword{tra}

