% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/poisson.tba.R
\name{poisson.tba}
\alias{poisson.tba}
\title{Poisson TBA as function of TRA and control mean}
\usage{
poisson.tba(tra, lamc)
}
\arguments{
\item{tra}{real number between 0 and 100}

\item{lamc}{real number in (0, Inf)}
}
\description{
Poisson TBA as function of TRA and control mean
}
\examples{
poisson.tba(80,2)
}
\keyword{pim}
\keyword{pip}
\keyword{tba}
\keyword{tra}

