% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/poisson.tra.R
\name{poisson.tra}
\alias{poisson.tra}
\title{Poisson TRA as function of TBA and control mean}
\usage{
poisson.tra(tba, lamc)
}
\arguments{
\item{tba}{real number between 0 and 100}

\item{lamc}{real number in (0, Inf)}
}
\description{
Poisson TRA as function of TBA and control mean
}
\examples{
poisson.tra(80,2)
}
\keyword{pim}
\keyword{pip}
\keyword{tba}
\keyword{tra}

