% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/zinb.tba.R
\name{zinb.tba}
\alias{zinb.tba}
\title{Zero-inflated negative binomial TBA as function of TRA, control mean, and theta}
\usage{
zinb.tba(tra, lamc, theta)
}
\arguments{
\item{tra}{real number between 0 and 100}

\item{lamc}{real number in (0, Inf)}
}
\description{
Zero-inflated negative binomial TBA as function of TRA, control mean, and theta
}
\examples{
poisson.tba(80,2)
}
\keyword{pim}
\keyword{pip}
\keyword{tba}
\keyword{tra}

