% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/zinb.theta.mom.R
\name{zinb.theta.mom}
\alias{zinb.theta.mom}
\title{Method of Moments estimator for Theta (dispersion parameter) in the Standardized Model-based Zero Inflated Negative Binomial model}
\usage{
zinb.theta.mom(yt.bar, yc.bar, s2t.hat, target.mean = 2)
}
\arguments{
\item{yt.bar}{the observed mean in the treatment group}

\item{yc.bar}{the observed mean in the control group}

\item{s2t.hat}{the observed sample variance in treatment group}

\item{target.mean}{the standardized control mean that is being targeted}
}
\description{
Method of Moments estimator for Theta (dispersion parameter) in the Standardized Model-based Zero Inflated Negative Binomial model
}
\examples{
zinb.theta.mom(yt.bar=0.85, yc.bar=1.15, s2t.hat=2.033, target.mean = 2)
}
\keyword{pim}
\keyword{pip}
\keyword{tba}
\keyword{tra}

